react-native fixes
