+++ packages/react-native/src/evm/components/Web3Button.tsx
@@ -141,6 +141,7 @@ export const Web3Button = <TAction extends ActionFn>({
       content = "Switch Network";
     } else if (
       actionMutation.isLoading ||
+      !contract ||
       connectionStatus === "connecting" ||
       connectionStatus === "unknown"
     ) {
